requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
certifi==2020.4.5.2
colorama==0.4.3
d6tcollect==1.0.5
d6tflow==0.1.9
decorator==4.4.2
defusedxml==0.6.0
docutils==0.16
entrypoints==0.3
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
lockfile==0.12.2
luigi==3.0.0
MarkupSafe==1.1.1
mistune==0.8.4
nb-conda==2.2.1
nb-conda-kernels==2.2.3
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.5
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
pip==20.1.1
prometheus-client==0.5.0
prompt-toolkit==3.0.5
pyarrow==0.17.1
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-daemon==2.2.4
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.1
scikit-learn==0.23.1
scipy==1.3.2
Send2Trash==1.5.0
setuptools==47.1.1.post20200529
six==1.15.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
tornado==5.1.1
traitlets==4.3.3
wcwidth==0.2.4
webencodings==0.5.1
wheel==0.34.2
wincertstore==0.2
zipp==3.1.0